move Alterable.java
